doc: work around groff bugs in html generation
groff version 1.22 and later has a bug in html generation (via -Thtml)
where .nf sections lose their indentation.  This makes the code examples
in our man pages all slammed together to the left margin.  Reverting to
groff version 1.18.1.4 avoids this bug, so the path to groff is changed
to /usr/local/bin for now to remind me to use the local install until
the bug is fixed.

Even with the old groff, blank lines in the code examples cause the
next line to have the incorrect indentation (off by a space).  Adding
a tab on the blank line avoids this bug while leaving the output of
the man command the same.

groff does document that the html driver is just beta quality, so
hopefully these issues will be fixed in future updates.