common: add VALGRIND_DO_PERSIST macro
In some cases valgrind may report dirty stores when we
don't care about it. Example:
If we would allocate an object, modify it and if the
transaction would be aborted there would be a dirty store.
We don't care about this store because in transaction abort
the object is deallocated. In such case we could use the
VALGRIND_DO_PERSIST macro in order to tell valring that everything
is ok with this dirty store.
We can't use VALGRIND_REMOVE_PMEM_MAPPING macro as we don't know
whether the transaction will be aborted or not.