obj: remove assert on pool_uuid_lo from list_insert_user
The pool_uuid_lo is not cleared when removing from list,
because the the object is OID_NULL if offset is zero. Thus
there cannot be an assert on pool_uuid_lo because it is
not possible to insert element to a list which the last
element was removed from.