common: replace parent UUID with poolset UUID
Pool set UUID shall not be equal to the UUID of any part of the pool set.
In particular, it's not the UUID of the first part of the pool set!
The same pool set UUID shall be used in all the replicas.  Also, this is
the one that is encoded in all the OID's.