obj: do not read uninitialized value on alloc
list_set_oid_redo_log has an optimization which avoids writing
pool_uuid_id into an oidp when it already has a correct value. However
on alloc oidp might be uninitialized, so we should not look at it. To
fix this, introduce a flag which tells whether we are allowed to look
at oidp and allow it from realloc/free/move, but disallow from alloc.

Found by valgrind/memcheck.