benchmarks: framework for benchmarks
This simple framework provides an interface for implementation
of various benchmarks. Each benchmark may be run with some common
arguments like: number of threads, data size, number of operations etc.
The results are presented in unified format as a CSV tables.
The framework allows to run scenarios defined in config file.