obj: fix bucket repopulate race
When multiple threads allocated memory linearly (like in a synthetic
benchmark) and encountered an empty default bucket, which triggered
the need to use another zone, the same memory blocks could have been
added to the bucket several times. This patch changes the logic so that
the bucket repopulation always happens safely behind a lock.