blk: use poolset_uuid as parent uuid for btt
Use the poolset_uuid as parent uuid for btt arena info instead of
the uuid field from pool header. It seems like the poolset uuid
is more appropriate field for this because if the replication
would be enabled for the pmemblk mode the other replica would
have different uuid fiel but the same parent_uuid.