common: fix overflow in replica/parts calculations
Both PART and REP are called with -1 as a 2nd argument.
The code assumes it will get the last part/replica, but for some values
of nreplicas/nparts it gets different values.
For example:
((unsigned)-1) % 7 == 3
Switching to signed types does not help:
-1 % 7 == -1
So the best fix is to avoid negative value by adding nreplicas/nparts
before modulo.