vmem: destroy pool locks
If a vmem pool is created using vmem_create_in_region() in a memory
region that was previously used by another vmem pool, it may happen that
some pthread locks used internally by jemalloc will be in the same
location as the locks used by the destroyed pool.  Since attempting to
initialize an already initialized lock results in undefined behavior,
it is important to destroy all the internal locks before the pool
is destroyed.