vmem, vmmalloc: align start of pool to 4MB
jemalloc divides passed memory between base allocator and "usable" space
in a way which depends on alignment of passed address.
It means the amount of usable space for files of the same size depends
on luck. This is quite suprising.

This patch takes the luck out and makes the division at fixed point.