test: fix obj_realloc read of unitialized values
util_checksum was called with size which was not a multiple of 4.
Additionally util.c was not compiled with DEBUG flag so assert
in util_checksum was not effective in catching this bug.
Even with -DDEBUG exact assert message was not visible because
out_init was not called.

Instead of fixing it, just provide (less optimized) checksumming
function in unit test framework and use it in obj_realloc.