obj: fix zalloc and zrealloc
This patch makes the zalloc an zrealloc functions to zero initialize
the entire allocated area - not just the requested area. The problem
was if the zallocated object was zreallocated, in such situation the
memory above requested area and below the usable size was not zeroed,
next in zrealloc the memory above the usable size was zeroed which
resulted in non-zeroed part of memory after zrealloc.

Ref: pmem/issues#120
Ref: pmem/issues#121