test: remove asserts from vmmalloc_init test
When library is compiled with default flags, malloc and __malloc_hook
do not point to the actual malloc function, but to the proper PLT entry.
And since this is the same PLT entry (both should point to malloc from
libvmmalloc), the pointers do match.
However, when library is compiled with -Bsymbolic-functions flag (which
is automatically added when building packages on some Linux distros),
__malloc_hook no longer points to PLT, but directly to libvmmalloc's
implementation of malloc.  So, even though the same function is
referenced in both cases, the pointers are different resulting
in assertion failure.