obj: refactor allocator buckets implementation
This patch does away with the getter functions from buckets and
instead the necessary fields are accessible directly from the
structure. It also abstracts away the memory blocks container so that
a tree-based one can coexist with a list-based, so that various
allocation policies (best-fit, first-fit, next-fit) can be a
compile-time or even run-time option.