benchmarks: fix evaluating size of a log pool
The sizes of pool header and pool descriptor were taken into account to
evaluate the size of a loog pool required to fulfill the benchmark
execution. The problam was that after opening the pool the size could
be lower because of page alignment requirement. This patch adds
additional 4k page size to the pool size.