test: fix makefile dependencies
When building in test directory with multiple jobs, the build sometimes
failed. There was the following error when linking pmemdetect tool
on one of its object files:

*.o: file not recognized: File truncated

The problem was that each and every test depends on pmemdetect tool and
there is a rule in src/test/Makefile.inc which builds the pmemdetect
from the test directory.

The tests and tools were building in parallel and thus somethimes the
pmemdetect was build twice (at least) and make could not synchronize
this because the rules were in different makefiles.

This patch fixes the dependencies in src/test/Makefile which fixes the
issue.