common: util.c refactoring
Move platform-specific utilities to util_platform.c.  When compiling
NVML for Linux or Windows, a platform-specific variant of this file will
be compiled/linked.

Files with platform-specific implementation are located in 'platform'
subdirectory.  The structure of this directory is the same as for the
main directory of the source tree.

This is a preparation for NVML for Windows support.