obj: fix 2 bugs in transaction abort
1) When application snapshotted an area from inside of
a PMEMmutex / PMEMrwlock, we would restore it on abort.
2) When application snapshotted 2 areas - one ending in the middle
of a lock and one starting in the middle of a lock, we would restore
the 2nd one.

Applications that snapshot part of a lock are buggy anyway, so for
correct apps this fix should not matter.

I noticed these bugs while cleaning up tx_restore_range function.