test: add windows UT support and multiple tests and VS projects
Also misc modifications to support windows compilation. Not all
tests are included in this patch, just the required framework,
two sets of tests, blk_nblock and blk_non_zero, and several but
not all VS project files for the tests.

Also addressed all compiler warnings and modified match files in
blk_nblock and blk_non_zero so they are OS independent. Win and
Linux will share match files except where output is so different
that it can't be matched reasonably. In those cases the test
and match files should end with a little "w" so they are
proprely matched with each other and are distinct from the
common sets.