win: address review comments
Also addressed all compiler warnings and modified match files in
blk_nblock and blk_non_zero so they are OS independent. Win and
Linux will share match files except where output is so different
that it can't be matched reasonably. In those cases the test
and match files should end with a little "w" so they are
proprely matched with each other and are distinct from the
common sets.