win: replace list routines with sys/queue.h routines
Previous list routines came from listentry.h included in the
Windows SDK and not something we want developers to have to
install.  Replace with BSD licensed routines in sys/queue.h
with a few extra wrappers to make VS happy.  Also a small
drive by on runtests.ps1 to make the output nicer.