win: move memory mapping routines to libpmem
Windows implementation of mmap(), munmap(), msync() and mprotect() has
been moved to libpmem.
On Windows, those functions are exported by libpmem, making them
available for all the other NVM libraries. So, there is no need to link
mmap_windows unit to each NVM library or test program.
libpmem now holds the only instance of file mapping list, so it is able
to find the file handle associated with the address passed to
pmem_msync() and flush the data to persistence, regardless of which
library created the file mapping.

Note: This is a temporary solution - a subject for change in the near
future.