vmem: fix multiple execution of configure script
Sometimes the configure script was invoked multiple times
for single build type (debug/nondebug). This was because the receipe
which invokes configure was associated with multiple targets
(files generated by configure) and the receipe was executed for
each target separately. This patch invokes the configure receipe for
single file (the first one) and creates a dependency on this file for
the rest of files generated by configure.