obj: simplify & optimize alloc class generation
The current default allocation classes result in duplicate buckets
supporting the same unit size because the bucket map for the class unit
size didn't point to the corresponding bucket. This patch addresses that
issue and also simplifies the related logic in a way that hopefully
makes it easier to reason about.