test: add windows tests for log_pool, add os specific err mapper
Add windows tests for log_pool and because windows doesn's support
all C errors, added a new abstraction os_strerror() that is a pass
through on Linux but on windows will map error strings unknown to
Windows based on the incoming code.

This includes a new linux platform.h for implementation of
abstracted OS calls.  The windows version is already there in
a windows specific include dir and included in all files vis
VS directive.  For linux the flie platform.h is included as
needed where those calls are used.