test: add windows test for log_pool
... and os specific err mapper

Add windows tests for log_pool and because windows doesn's support
all C errors, added a new abstraction util_strerror() that is a pass
through on Linux but on windows will map error strings unknown to
Windows based on the incoming code. The unit test equivalent is
called ut_strerror() and implemented so that the UT framework need
not rely on libutil.

This includes new util_linux.c and util_windows.c files to hold
utility like OS abstractions. The util_strerror() function is dup'd
in the unit test code until we implement mock functions for win.