Use movnt by default
The CPU dispatching code checking for SSE2 is wrong,
and very misleading.
The code is either compiled using SSE2 instructions - in which
case, such instructions might end up in other places in the binary -
or compiled without such instructions, in which case the
_mm_stream_si128 identifier is not present, thus doesn't compile.
The right way to do such CPU feature based dispatching is to
place appropriate code segments in different translation units,
compiled with appropriate flags e.g. -msse2
and to conditionally call such code from translation units compiled
with some least common denominator platform in mind.

The code not using movnt is left in the code, to available via
an environment variable, as before.