obj: allow for memblocks /w size index equal to 64
The size index of a run memory allocation is currently capped by
RUN_UNIT_MAX compile-time constant, which happens to be 8. This caused
a potential issue to creep into the algorithm. The issue could trigger
undefined behavior to present itself when the above mentioned constant
is changed to its edge value of 64.