common: fix isblank usage in set.c
According to ISO:
"the argument is an int, the value of which shall be representable
as an unsigned char or shall equal the value of the macro EOF."

Thus a negative char value supplied to isblank ( except EOF ) is undefined
behaviour. E.g. the implementation might use a lookup table, in which
case a negative index can result in a crash.