win: fix bug in RUNTESTS
Fix incorrect implementation of -s.  -s should specify the test
script name but was being used as test dir name.  Test dir name
on linux does not require a switch however in PS we can't have
a combo of switched args and positional args so the fix is to
use -i to specify test dir.  So, for example, to run just TEST1
in the blk_nblock test dir:

.\RUNTESTS -i blk_nblock -s TEST0

to run all test in the dir:

.\RUNTESTS -i blk-nblock