test: introduce a new way of setting up tests requirements
The new way of setting a test requirements is by using a config.sh file located
in a test sub-directory. The config.sh file structure is described in
src/test/config.sh_example. It is the recommended method but it is still under
development and does not cover all possible requirements. The old way of setting
test requirements using require_* functions is still valid and can be mixed with
the new way.