cpp: introduce the allocator
Add the persistent memory allocator. Modified makefiles to enable
custom libsdtc++ and libc++ installations. Add basic tests, but
the container tests. Then need a slightly modified version of
libc++ to be able to compile.

The allocator introduces the persistency_type typedef, which is
used by the modified libc++ implementation for appropriate
container metadata handling.

This commit does not include Windows related projects for the
unit tests. That is because VS2015 does not handle allocators
with fancy pointers correctly. The debug version of the containers
will most certainly segfault on these tests. It is probable
that VS2017 will handle this a lot better.

Linux tests are available, however they are not compiled on travis
yet, because a non-standard, modified version of libc++ is needed.