obj: fix benign race in the huge chunk free logic
The free code path includes a coalescing step that searches for the
neighbouring blocks by traversing the on-media layout and then it verifies
the ownership of the block by looking it up in the bucket tree. Due to the
fact that the only lock during that process is the bucket mutex, the layout
was left unprotected. In general that algorithm was correct but it produced
errors in helgrind, so for correctness sake from now on the on-media layout
is protected by the default bucket lock. This should have minimal impact
on scaling but will ensure that the thread error detecting tools will be
happy.