rpmem: fix race in obj_rpmem_heap_interrupt
The obj_rpmem_heap_interrupt test does not call rpmem_close function,
which means the pool is not closed gracefully and synchronously. In the
TEST0 script the second rpmem connection is established immediately on
the same pool. This is a race because the previous connection may still
hold the file lock when trying to open the pool in the second
connection. In this patch the pool is removed and re-created using
pmempool before the second connection.