rpmem: fix closing rpmemd on remote node
In one of previous patches (21adbdc5c9dea710293c3586a63dc57588c3b61b)
the allocation of pseudo terminal for ssh session has been disabled in
order to make sure the session is transparent for binary data.

However disabling allocation of pseudo-terminal introduced an issue
with closing the rpmemd process on remote node. If there is no pseudo
terminal allocated the sshd doesn't send SIGHUP signal to the rpmemd
and thus the rpmemd process hangs in some situations if we terminate
the ssh on initiator node. The hang occurs when remote node is waiting
for in-band connection but the client has already terminated.

This fix changes a little bit a flow of establishing and terminating
the in-band connection:

- Waiting for in-band connection is now spawned in background
  thread when the main thread is reading from stdin - in this case we
  know when the client side has terminated.
- Removing pool if rpmem_create fails is done by rpmemd not librpmem.
- Timeout for accepting connection.
- Some error handling fixes.

Ref: pmem/issues#235