obj: simplify palloc locking scheme
This patch reduces the number of locks acquired and released in the
duration of the palloc_operation function and also removes the recursive
attribute from the run locks. This also fixes a very rare race in which
threads freeing blocks from the same run tried to concurrently degrade
the same run.