libpmempool: refactor pool_params_parse()
The clang analyzer reports a 'Function call argument
is an uninitialized value' warning claiming that stat_buf is not
initialized. This is false positive because the stat_buf is initialized
when params->is_poolset is false and is_btt is false. This patch
refactors a little bit the function because the flow was too
complicated and it silences the clang analyzer at the same time.

Ref: pmem/issues#312