obj: rewrite palloc free and unused run handling logic
This patch introduces deferred updates of the heap's transient state in the
case of free. Currently each free performs multiple tree lookups to locate
neighbours, then performs a tree insert and then finally checks if the
run has been completed freed and if that's the case it removes relevant
entries from the tree and changes the run back to a chunk. All of this is
very expensive and requires a lot of locks in order to be performed safely.
The new changes reduce the free to a simple bit flip under a single lock,
and the transient container update is performed only if the current zone
exhausted all of its memory.

This is a huge simplification of the allocator's algorithm which allows for
reduced locking and more straightforward looking code.