libpmempool: sync, transform API as experimental
We would like to add a support for reporting progress for both
pmempool_sync() and pmempool_transform(). It means we probably will
need to change the API significantly. Thus it is reasonable to mark
both functions as EXPERIMENTAL so users won't stick to the API. We have
pmempool sync and trasform commands which are stable and this seems to
be enough to provide a solution for recovery and pool sets management
functionality.