common: integrate Coverity scans with Travis
The new travis.yml introduces a new entry/job in the build matrix to
trigger the Coverity Scan build for static code analysis.
The "COVERITY" job is executed only if the build is made from
"coverity_scan" branch and skipped (early exit) otherwise.
On the other hand, the normal build jobs are skipped for "coverity_scan"
branch.
Thanks to that, we can have a single travis.yml file on all the branches
and simply push the code from "master" to "coverity_scan" periodically
(i.e. daily) to trigger Coverity scans.