test: suppress helgrind issues /w adjacent blocks
This is an on-purpose race that has no negative effects. In happens during
a free of a huge block, when the algorithm tries to coalesce the block with
its free neighbours. The blocks are first identified in the persistent
layout without grabing a lock and then they are retrieved from a protected
container - this ensures that the block is correct. Even if the data that
was read is stale or invalid, nothing bad can happen - the chunk that is
being freed simply won't get coalesced.