obj: fix pmemcheck for large free undo logs
For small tx undo logs, the vector structure uses an embedded array that is
located in lane, if the undo log needs to grow beyond the initial size, it
allocates an array from a heap. The issue was that the lanes are added
to the pmemcheck tx ignore list whereas the heap allocations must be
manually added to the transaction to suppress the pmemcheck warning.