common: add VS shared property sheets
Adds shared property sheet files for NVM libraries VS projects.

The properties that are common for all the libraries (include paths,
preprocessor macros, optimization settings, etc.) can be stored in
one file (actually two files - for debug and release configurations),
so if there is a need to i.e. add/modify some include path in all
the projects, it can be done by modification of the .props files.
No need to modify all the *.vcxproj files anymore.

Most of the properties in the library project files have their default
values, as all the custom settings are now in *.props files.
Only libvmem has some additional include paths and preprocessor macros
(for jemalloc), and those are still in libvmem.vcxproj.