pmempool: rm command update
- don't remove a local poolset file by default
- add options for specyfying which poolset files shall be removed
  (local, remote or both)
- update documentation
- use rpmem_remove API

This patch changes the default behavior of `pmempool rm` because it
used to remove the local poolset file by default. However this change
is in the (hopefully) less aggressive direction as it now removes less
than it used to before, thus this shall not be a big problem for any
users to notice the change.