test: add extra offset field to stacktraces
The new field is the address the instruction pointer points
to inside the object file. This is an offset relative to start
of an object, which might a shared object loaded to an offset
affected by ASLR, thus without dumping such extra information
it is difficult to determine where in such a shared object a
stack frame is.

This offset printed can be used with the addr2line tool, e.g.:

addr2line -a 0x123abc -e src/debug/libsomething.so.1.0.0.0

In the above example, the actual value of RIP would be

(address where .so is loaded) + 0x123abc

which obviously can't used with addr2line out of the box.