obj: add instrumentation to all instances of new chunks
It's a third attempt at fixing helgrind, but this patch ultimately gets rid
of all the false positives.

This change makes it so that the helgrind is instrumented to forget all the
state about chunks that are inserted into the free chunks tree. This means
that there won't be any false-positives about conflicts with changes that
are old, unrelated and possibliy not protected by a lock.