vmem: fix reallocation to smaller size
This is another variant of the fix introduced in commit cd868a1.
The new version does not break the Valgrind instrumentation and jemalloc
stats.

If shrinking the huge allocation (>=4MB) to some smaller size fails due
to the lack of memory for a new arena, return the original allocation
pointer instead of NULL.
For allocations larger than a single chunk, try to shrink them to a
single chunk first, then there is a chance that the actual reallocation
will succeed.

Ref: pmem/issues#24