common: use rpath when linking libraries
If user installs NVML to $HOME and try to build his application using
pmemobj the linking phase fails, because of missing path to libpmem.
Solve it by setting runtime path of the directory with libpmem for all
libraries.
In general use of rpath is discouraged for distro packages, so disable
it for both Debian and Fedora packages.