test: simplify Unicode support in test framework
In general, all the tests are run with Unicode paths, both on Linux and
Windows.  A suffix containing non-ASCII characters is appended to the
test directory path generated for each test case.

On Windows, most of the tests are not compiled with Unicode support
(there is no _wmain() function), so to make them work with Unicode
paths/strings, we need to obtain the original, UTF16-encoded command
line arguments and convert them to UTF-8.  This is done in START/DONE
macros - completely transparent to the rest of the test code.

Since NVML examples do not use our test framework (START/DONE macros),
the ex_* are the only tests that cannot be run on Unicode paths.
In such case, the mentioned suffix is not appended to the test directory
path.