test: introduce OS abstraction layer
Similar to the NVM libraries, our test framework should also use
OS abstraction layer, instead of calling the system functions directly.
This guarantees the code of all the tests is portable and supports
UTF8-encoded arguments.

The OS-abstraction layer code is located in src/common for now, but since
it's also used by libut now, it should be moved to another location,
i.e. "src/os" (replacing "src/windows"), but this is a subject for
a separate patch.