obj: add replicas to a pmemcheck ignore list
Currently the instrumentation for pmemcheck is done only for main
mapping, and all of the replicas are effectively ignored. This leads to
an issue in which none of the pmemcheck tests pass when using replicated
poolsets. This could be solved in several different ways, including
changing our instrumentation so that it propagates to replicas, but that
would a) be very expensive in terms of CPU time and b) wouldn't really
test anything. After some deliberation, we've (me and @tkapela) decided
ignore the memory that belongs to local replicas, because it's the
cheapest and simplest solution we've came up with.