test: limit number of arenas in vmem_stats
By default, jemalloc allocates 4 * num_of_cpu arenas. If statistics
are enabled, it also allocates the corresponding per-arena stat
structures for all the arenas, even if only one arena is actually used
(single-threaded program).  If this test is run on a machine with
the number of cores >=64, the amount of memory required to keep all
the statistics is larger than the usable space of the pool.

The easiest workaround is to limit the number of arenas for this test
by setting JE_VMEM_MALLOC_CONF environment variable.  Thus, the test can
be run on any machine and we don't need to calculate the pool size
based on number of available CPU cores.

Backported from master (ccbf4cb).