test: increase runtest.ps1 performace
This has been done by reducing amount of process spawned by RUNTEST.ps1.
RUNTEST.PS1 now parses test script to detect in which
configurations should run that test. Old way was to run test in every
posible configuration<spawn process there> and test
decides if he want to be run in this configuration or not, although
this model isn't a problem on linux, as spawning a new process is
very fast<fork>, on windows this is very very slow.