test: improve runtests.ps1 performance
This is done by reducing the number of processes spawned
by RUNTESTS.PS1, as spawning a new process on Windows is very costly.

Now, before executing given test, RUNTESTS.PS1 first parses TESTx.PS1
script to determine for which configurations it should be run, and only
for those configurations the test is actually executed (the new process
is spawned).

The original approach (still used on Linux) was to run given test
script (new process) for each posible configuration and make a decision
on whether to continue the test for given configuration or not inside
the TESTx script.