examples: restructure libart directory: libart3
Create new subdirectory libart3. This contains the 3rd port
of libart to libpmemobj. The key idea of this port is to
use different RAM types for
   - leaf nodes: are kept in PMEM
   - tree nodes: are in DRAM
This approach is similar to pmemkv.

The expectation is that read operations
(lookup, tree iteration) are faster with this approach.

Adjust PROGS and LIBRARIES in Makefile of all libart variants.