common: add env var for skipping expensive checks
Currently it's not possible to profile applications using pmemobj.
Release versions of the library don't have debug symbols and
debug versions are very slow because of single check in
pmemobj_tx_commit.

To allow profiling introduce PMEM_DEBUG_DISABLED_CHECKS_MASK
environment variable which is a bitmask of checks user can disable
and define "1" as bit that disables the check in pmemobj_tx_commit.