common: set some flags for release packages
-ggdb does not affect generated code and for distro packages all
debug info is moved to a separate package with debug informations.

-fno-omit-frame-pointer affects generated code, but its overhead
should be negligible.

With those two flags in release builds it's possible to analyze
performance traces. Without -fno-omit-frame-pointer the stacktraces
are flat and spread randomly making the analysis impossible.
Without -ggdb stacktraces are reliable, but the part in NVML has
"unknown" as function names. (-ggdb without -fno-omit-frame-pointer
is useless)