pmem: (win) fix PMEM_MMAP_HINT implementation
Fixes calculation of mapping hint address when the address specified
via PMEM_MMAP_HINT is busy (i.e. when more than one pool is open
by the process).

Originally, on Windows the library would fail to open/map the pool
in such case.  With the fix it would behave similarly to linux version
and it would scan the process adress space for the next unused region
above the hint address that is large enough to hold the entire pool.