test: (win) fix recording open files
Adds missing "break" that caused all the handles to be actually
ignored.

Remove querying the handle for object/file name - only the object
type name is recorded.  In some cases NtQueryObject may hang
the process, and there is no way to handle that other than performing
the query from another thread, and killing it if it stops to respond.