obj: change alloc policy to allocate lowest-first
This patch changes the seglists container to work in a FIFO fashion,
and in turn, return the memory blocks with the lowest offset - the
ones that were added first.

This patch shouldn't affect the overal throughput of synthetic
benchmarks, but might improve locality of data of some workloads
(e.g., creating a linked list).