obj: move run info from alloc classes
This patch slightly refactors the allocation classes so that the
information about a run (e.g., bitmap size, nallocs) are in a
separate structure. This allows the rest of the code to operate
on run memory blocks without requiring an equivalent allocation
class.

This fixes issues where memory blocks created using old allocation
classes were incorrectly interpreted due to usage of a wrong class
information.