obj: fix split snapshot being cut by 16 bytes
If there's not enough space in a tx cache, the snapshot is split into
two parts: one part is small enough to fit in the remaining space in the
cache, and the second one lands in a new cache.
There is a bug in this logic that results in the second part having
both offset and size cut by 16 bytes. This patch fixes that and adds
a test for that logic.