test: fix add_range_direct under pmemcheck
There seems to be a problem with valgrind pmemcheck not falling back
to using anonymous mmap when sbrk fails. This leads to volatile
malloc failures when a test exceeds the applications break. What's
worse, the behavior depends on the operation system malloc
implementation and so it's pretty difficult to predict. For now, I've
reduced the amount of allocated memory for the problematic test but
we will probably need to tackle that problem eventually.