obj: optimize thread utilization of buckets
This patch improves the way buckets are assigned by counting how many
threads are currently using a given bucket and always picking the
first one that is the least used.
This patch also changes the 'bucket_cache' name to a more meaningful
'arena', a term that is widely used and understood.