common: change pool size validation
Do not artifically set lower bound for a pool part size. A pool part
has to be big enough to incorporate a pool header. If the pool type
requires certain minimal pool size this limit is enforced on a net
pool size not on sizes of each part separately. The net pool size is
described in each pool type documentation.