make: allow the objcopy and pkg-config to be overridden
When building nvml under different environments (ie. buildroot),
the objcopy and pkg-config need to be settable thru the environment
since the gcc, binutils, etc might not match the host during this
build.

This prevents errors on relocation symbols not being understood
by link specically discovered when working on Ubuntu 14.04 and
build 2017.02 to add an nvml package support.

Signed-off-by: Charles Hardin <ckhardin@exablox.com>