benchmark: add memcpy operation to blk
Adds an additional operation type that allows to compare performance
of pmemblk_read/write vs. pmem_memcpy.  Replaces 'file-io' argument
with 'operation', which can be 'blk', 'file' or 'memcpy'.

Replaces 'random' flag with 'mode' argument, which controls
the read/write pattern amd culd be 'stat', 'seq' or 'rand'.

For tests run on Device DAX, 'files-size' argument can be used
to limit the size of the device that is actually used for the test.