obj: fix potential deadlock during realloc
When two threads simultaneously tried to reallocate two objects,
and when the source size of the object A matched the destination size
of the object B, and the the object A resided on the same run as
object B, there was a chance for a deadlock. The only solution to
this problem while keeping atomicity guarantees (we have to keep the
redo log for the entire operation), is to sort locks by their address,
and that's what this patch does.

Found by helgrind.