obj: fix same-size realloc volatile state change
Currently, when performing realloc, the new block is unconditionally
allocated, and when the decision to bail out (because the request is
a no-op) happens, the volatile state changes are not reversed. This
is a pointless waste of CPU cycles, and in some cases might even lead
to persistent memory being locked out until the next heap incarnation.