rpmem: fix setting errno if ssh failed
There is not any good solution to propagate an error from external
process to the errno variable. In case of ssh if the ssh failed it
returns 255, otherwise it returns the command's status code but
assigning this to errno doesn't make any sense. The most probable
situation is that there is something wrong with the arguments to the
ssh or rpmemd, thus this patch sets EINVAL in case of ssh failure.

Ref: pmem/issues#620