test: improve vmmalloc_malloc_hooks test
Adds explicit malloc hooks initialization instead of using
__malloc_initialize_hook variable.  The __malloc_initialize_hook
variable is no longer exported by the recent versions of glibc.

It also adds one more test scenario without libvmmalloc pre-loaded,
so the hooks are expected to be called.  This helps to detect the case
when the hooks are not initialized properly, or no longer supported by
glibc.