obj: add early evaluation of runs free space
Evaluation of the free space available in the runs is done lazily and
very infrequently - effectively only when the algorithm hits an OOM
condition after exhausting the memory present in the volatile state.

The described mechanism works well with very large workloads (16GB++),
but produces very large external fragmentation otherwise.

This patch introduces a compromise. The free memory in the runs will
be recalculated after the user has freed about two times the size
of a single run for the allocation class. This means that the heap
will no longer create new runs when more than a full run of memory is
already available.
This unfortunately comes at a cost in workloads that perform a lot of
deallocations at a regular basis, but in most scenarios,
the performance penalty will be mostly amortized by other heap
operations.