common: add support for headerless pools
The prototype support for pool sets with the pool headers only in the
first part of each replica.  It is required to allow creating pools
that span multiple Device DAX devices with internal alignment other than
4KiB.  Of course, it can be used with pool sets built on regular files,
as well.
When this option is enabled, it must be set for all the replicas
in the pool set (including remote ones).

In this prototype, the new option must enabled by setting environment
variable "PMEM_NO_HDRS=1".  Eventually, it will be integrated with
pool set options (see PR #2268).

TBD:
- add more tests
- tests for remote replicas on 2M Device DAX devices
- backward compatibility tests (old libs cannot open "headerless pool")
- pmempool info/dump/... - print info about compat flags, etc.
- pmempool transform - convert existing pool set to "headerless" pool
  or vice versa + tests
- man pages update

OPENS:
- Since prev/next part UUID does not make sense in case of "headerless"
  pool, we could add the number of parts / total pool size to
  the pool header, which would allow to verify the data in pool set
  file and/or to detect missing parts.