common: add support for headerless pools
Adds the initial support for pool sets with the pool headers only in the
first part of each replica.  This feature is required for creating pools
that span multiple Device DAX devices with internal alignment other than
4KiB.  Of course, it can be used with pool sets built on regular files,
as well.
When this option is enabled, it must be set for all the replicas
in the pool set (including remote ones).